DynamoDBスキーマの内容をDTOに寄せた
