+++ packages/serverless/src/dynamo-db/casual-match-entries.js
@@ -1,15 +1,17 @@
 // @flow
 
 import {DynamoDB} from "aws-sdk";
+import type {UserID} from "../dto/user";
+import type {ArmDozerId, PilotId} from "gbraver-burst-core";
 
 /** casual_match_entriesのスキーマ */
-type CasualMatchEntry = {
+type CasualMatchEntriesSchema = {
   /** ユーザID */
-  userID: string,
+  userID: UserID,
   /** 選択したアームドーザのID */
-  armdozerId: string,
+  armdozerId: ArmDozerId,
   /** 選択したパイロットのID */
-  pilotId: string,
+  pilotId: PilotId,
   /** コネクションID */
   connectionID: string,
 }
@@ -36,7 +38,7 @@ export class CasualMatchEntries {
    * @param entry 追加する項目
    * @return 処理が完了したら発火するPromise
    */
-  put(entry: CasualMatchEntry): Promise<void> {
+  put(entry: CasualMatchEntriesSchema): Promise<void> {
     return this._client
       .put({TableName: this._tableName, Item: entry})
       .promise();
@@ -47,7 +49,7 @@ export class CasualMatchEntries {
    *
    * @return 取得結果
    */
-  async scan(): Promise<CasualMatchEntry[]> {
+  async scan(): Promise<CasualMatchEntriesSchema[]> {
     const resp = await this._client
       .scan({TableName: this._tableName, Select: "ALL_ATTRIBUTES"})
       .promise();
